refactor(ui): ReminderSectionをReminderContentに名称変更し構造を改善
- ReminderSectionをReminderContentに名称変更
- ReminderListの機能をReminderContentに統合
- 階層構造をフラット化し、責務を明確化
- HomeScreen内でのコンポーネント名を分かりやすく整理
